Fix file copy

+++ cmd/cli/cmd/scan.go
@@ -269,10 +269,10 @@ func persistDirNode(dn *DirNode, pid int64, sid int64, db *sql.DB, rp string) {
 			var id int64 = 0
 			db.QueryRow("SELECT `id` FROM `nfile` WHERE `hash` = ?", f.Hash).Scan(&id)
 			if id != 0 {
+				// its repeated but from previous scans, already in the database
 				f.ID = id
 				f.FileExists = true
 			} else {
-				// its repeated but from previous scans, already in the database
 				res, err := stmtFile.Exec(f.Name, f.Extension, strings.Replace(f.Path, rp, "", 1), f.DateModified, f.Size, f.Hash)
 				if err != nil {
 					fmt.Printf("error inserting nfile: %v\n", f)
@@ -353,31 +353,36 @@ func getSourceDirectories() []DirItem {
 }
 
 func copyFiles(s *Scan, np string) *Scan {
+	fmt.Println("")
 	for i, f := range s.files {
-		fmt.Printf("[%v/%v] copying file: '%v'\n", i, len(s.files), f.Name)
-		rp := strings.Replace(f.Path, s.root.info.Path, "", 1)
-		fp := path.Join(np, rp)
-		err := os.MkdirAll(fp, 0755)
-		if err != nil {
-			fmt.Printf("failed to create directory: '%v' - %v\n", fp, err)
-		} else {
-			i, err := os.Open(f.Path)
-			if err == nil {
-				defer i.Close()
-				o, err := os.Create(fp)
-				if err == nil {
-					defer o.Close()
-					_, err = io.Copy(o, i)
-					if err != nil {
-						fmt.Printf("failed to copy file: '%v' to '%v' - %v", f.Name, fp, err)
-					}
-				} else {
-					fmt.Printf("failed to create file to copy: %v - %v", fp, err)
-				}
-
+		if !f.FileExists {
+			fmt.Printf("[%v/%v] copying file: '%v'\n", i, len(s.files), f.Name)
+			rp := strings.Replace(f.Path, s.root.info.Path, "", 1)
+			fp := path.Join(np, rp)
+			err := os.MkdirAll(filepath.Dir(fp), 0755)
+			if err != nil {
+				fmt.Printf("failed to create directory: '%v' - %v\n", fp, err)
 			} else {
-				fmt.Printf("failed to open file to copy: %v - %v", f.Path, err)
+				i, err := os.Open(f.Path)
+				if err == nil {
+					defer i.Close()
+					o, err := os.Create(fp)
+					if err == nil {
+						defer o.Close()
+						_, err = io.Copy(o, i)
+						if err != nil {
+							fmt.Printf("failed to copy file: '%v' to '%v' - %v", f.Name, fp, err)
+						}
+					} else {
+						fmt.Printf("failed to create file to copy: %v - %v", fp, err)
+					}
+
+				} else {
+					fmt.Printf("failed to open file to copy: %v - %v", f.Path, err)
+				}
 			}
+		} else {
+			fmt.Printf("[%v/%v] skipping existing file: '%v'\n", i, len(s.files), f.Name)
 		}
 	}
 
@@ -390,7 +395,7 @@ func scan(ccmd *cobra.Command, args []string) {
 	sd := getSourceDirectories()
 
 	sp := viper.GetString("scan_path")
-	//np := viper.GetString("nostalgia_path")
+	np := viper.GetString("nostalgia_path")
 	fmt.Printf("\nscan_path: %v\ntags: %v\nsource: %v\n", sp, strings.Join(strings.Split(tags, ","), ","), source)
 	fmt.Println("source directories:")
 	var sourceID int64
@@ -452,15 +457,15 @@ func scan(ccmd *cobra.Command, args []string) {
 
 	// persist changes
 	partial = time.Now()
-	fmt.Printf("persist changes...")
+	fmt.Printf("\npersist changes...")
 	_ = persist(s)
 	elapsedpartial = time.Since(partial)
 	fmt.Printf("OK (%v)\n", elapsedpartial)
 
 	// copy files
-	// partial = time.Now()
-	// fmt.Printf("copy files...")
-	// _ = copyFiles(s, np)
-	// elapsedpartial = time.Since(partial)
-	// fmt.Printf("OK (%v)\n", elapsedpartial)
+	partial = time.Now()
+	fmt.Printf("\ncopy files...")
+	_ = copyFiles(s, np)
+	elapsedpartial = time.Since(partial)
+	fmt.Printf("OK (%v)\n", elapsedpartial)
 }
